Fix Phase 1 build errors: missing using directive and uint cast
- Add `using HistorianSyncTool.UI;` to MainForm.cs (fixes 8x CS0103 AppTheme not found)
- Cast count to uint in InterpolatedQuery call (fixes CS1503 int/uint mismatch)

+++ Services/HistorianDataService.cs
@@ -57,7 +57,7 @@ namespace HistorianSyncTool.Services
         {
             return Retry(() =>
             {
-                DataQueryParams query = new InterpolatedQuery(from, to, count, tagName)
+                DataQueryParams query = new InterpolatedQuery(from, to, (uint)count, tagName)
                 {
                     Fields = DataFields.Time | DataFields.Value | DataFields.Quality
                 };
